don't enforce an order on scopes: instead check if order is provided
this broke the use of some builtin laravel scopes, like latest() and oldest(), since the default 'asc' order was passed as the 2nd parameter to the scope, and wrongfully interpreted.

it is still possible to pass a custom order to a scope, it just needs to be provided as an array, instead of just a string.